kpi-engine/requirements.txt
pandas
python-dotenv
sympy
requests
fastapi[standard]==0.109.0
uvicorn[standard]==0.27.0